# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Immoware Hub API",
        "version": "v1",
        "summary": "REST-API des Immoware Hub für Spiegeldaten aus Immoware24, Hub-Vorgänge, Verzeichnis und Health.",
...
        "contact": {
            "name": "Hausverwaltung Müller GmbH"
        }
    },
    "servers": [
        {
            "url": "https://immoware.muellerhv.de",
            "description": "Produktion"
        }
    ],
...
    "paths": {
        "/api/docs": {
            "get": {
                "operationId": "api_docs_html",
                "summary": "API-Dokumentation",
                "tags": [
                    "docs"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "HTML-Dokumentation"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    }
                },
                "x-scope": [],
                "x-effect": "hub",
                "x-evidence-status": null,
                "x-phase": 1
            }
        },
...
        "/api/v1/documents": {
            "get": {
                "operationId": "api_v1_documents_index",
                "summary": "documents auflisten",
                "tags": [
                    "documents"
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/page"
                    },
                    {
                        "$ref": "#/components/parameters/per_page"
                    },
                    {
                        "$ref": "#/components/parameters/cursor"
                    },
                    {
                        "$ref": "#/components/parameters/sort"
                    },
                    {
                        "$ref": "#/components/parameters/fields"
                    },
                    {
                        "$ref": "#/components/parameters/q"
                    },
                    {
                        "$ref": "#/components/parameters/updated_since"
                    },
                    {
                        "$ref": "#/components/parameters/external_id"
                    },
                    {
                        "$ref": "#/components/parameters/status"
                    },
                    {
                        "$ref": "#/components/parameters/include_deleted"
                    },
                    {
                        "name": "property_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "unit_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "contact_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "case_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "folder_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "content_type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "document_type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "origin",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "429": {
                        "$ref": "#/components/responses/RateLimited"
                    },
                    "503": {
                        "$ref": "#/components/responses/ServiceUnavailable"
                    },
                    "200": {
                        "description": "Seite",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Document"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PageMeta"
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/Links"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    }
                },
                "x-scope": [
                    "documents:read"
                ],
                "x-effect": "hub",
                "x-evidence-status": "VERIFIZIERT",
                "x-phase": 1
            },
            "post": {
                "operationId": "api_v1_documents_upload",
                "summary": "Upload-Antrag für den Posteingang anlegen",
                "tags": [
                    "documents"
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/IdempotencyKey"
                    }
                ],
                "responses": {
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "429": {
                        "$ref": "#/components/responses/RateLimited"
                    },
                    "503": {
                        "$ref": "#/components/responses/ServiceUnavailable"
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "409": {
                        "$ref": "#/components/responses/Conflict"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationFailed"
                    },
                    "202": {
                        "description": "Upload-Antrag angenommen",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/DocumentUploadResult"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "501": {
                        "$ref": "#/components/responses/NotImplemented"
                    }
                },
                "x-scope": [
                    "documents:write"
                ],
                "x-effect": "immoware24",
                "x-evidence-status": "VERIFIZIERT",
                "x-phase": 1,
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/DocumentUpload"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/documents/uploads/{uuid}": {
            "get": {
                "operationId": "api_v1_documents_uploads_show",
                "summary": "Status eines Upload-Antrags (write_operation) lesen",
                "tags": [
                    "meta"
                ],
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "429": {
                        "$ref": "#/components/responses/RateLimited"
                    },
                    "503": {
                        "$ref": "#/components/responses/ServiceUnavailable"
                    },
                    "200": {
                        "description": "Status des Upload-Antrags",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/DocumentUploadResult"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "x-scope": [
                    "documents:write",
                    "sync:read"
                ],
                "x-effect": "hub",
                "x-evidence-status": "HUB",
                "x-phase": 1
            }
        },
        "/api/v1/documents/{id}": {
            "get": {
                "operationId": "api_v1_documents_show",
                "summary": "documents einzeln lesen",
                "tags": [
                    "documents"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "429": {
                        "$ref": "#/components/responses/RateLimited"
                    },
                    "503": {
                        "$ref": "#/components/responses/ServiceUnavailable"
                    },
                    "200": {
                        "description": "Einzelressource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Document"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "x-scope": [
                    "documents:read"
                ],
                "x-effect": "hub",
                "x-evidence-status": "VERIFIZIERT",
                "x-phase": 1
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "apiKey": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "hub_live_<prefix>_<secret>",
                "description": "API-Key im Authorization-Header. Scopes je Key, siehe x-scope je Operation. Der Scope admin umfasst alle Scopes."
            }
        },
        "parameters": {
            "page": {
                "name": "page",
                "in": "query",
                "schema": {
                    "type": "integer",
                    "minimum": 1,
                    "default": 1
                }
            },
            "per_page": {
                "name": "per_page",
                "in": "query",
                "schema": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 500,
                    "default": 100
                },
                "description": "Maximal 500. Gilt für Offset- und Cursor-Pagination."
            },
            "cursor": {
                "name": "cursor",
                "in": "query",
                "schema": {
                    "type": "string"
                },
                "allowEmptyValue": true,
                "description": "Cursor-Pagination statt page: leer für die erste Seite, danach meta.next_cursor. Opaque (Base64url-JSON aus Sortierschlüssel und id), an sort und Filter gebunden; ungültig liefert 400 invalid_cursor. Antwort ohne total, mit meta.next_cursor und meta.prev_cursor."
            },
            "sort": {
                "name": "sort",
                "in": "query",
                "schema": {
                    "type": "string",
                    "default": "-updated_at"
                },
                "description": "Feld, optional mit führendem Minus für absteigend. Erlaubte Felder je Ressource."
            },
            "fields": {
                "name": "fields",
                "in": "query",
                "schema": {
                    "type": "string"
                },
                "description": "Sparse Fieldsets, kommagetrennt. id und provenance werden immer ausgegeben."
            },
            "q": {
                "name": "q",
                "in": "query",
                "schema": {
                    "type": "string"
                },
                "description": "Volltextsuche über die suchbaren Spalten der Ressource."
            },
            "updated_since": {
                "name": "updated_since",
                "in": "query",
                "schema": {
                    "type": "string",
                    "format": "date-time"
                },
                "description": "Änderung im Hub seit diesem Zeitpunkt (Hub-updated_at, nicht Immoware24)."
            },
            "external_id": {
                "name": "external_id",
                "in": "query",
                "schema": {
                    "type": "string"
                },
                "description": "Exakter externer Schlüssel des Quellsystems."
            },
            "status": {
                "name": "status",
                "in": "query",
                "schema": {
                    "type": "string"
                }
            },
            "include_deleted": {
                "name": "include_deleted",
                "in": "query",
                "schema": {
                    "type": "boolean",
                    "default": false
                },
                "description": "Soft-gelöschte Spiegeldaten einbeziehen."
            },
            "IdempotencyKey": {
                "name": "Idempotency-Key",
                "in": "header",
                "required": true,
                "schema": {
                    "type": "string",
                    "maxLength": 128
                },
                "description": "Pflicht für POST und PATCH. Wiederholung liefert die gespeicherte Antwort (24 Stunden) mit Header Idempotent-Replayed: true; abweichender Body liefert 409 idempotency_mismatch."
            }
        },
        "headers": {
...
            "Retry-After": {
                "schema": {
                    "type": "integer"
                },
                "description": "Sekunden bis zum nächsten Versuch (bei 429)."
            },
...
        },
        "responses": {
            "BadRequest": {
                "description": "400 bad_request, invalid_filter, invalid_cursor, idempotency_key_required",
                "content": {
                    "application/problem+json": {
                        "schema": {
                            "$ref": "#/components/schemas/Problem"
                        }
                    }
                }
            },
            "Unauthenticated": {
                "description": "401 unauthenticated, key_expired, key_revoked",
                "headers": {
                    "WWW-Authenticate": {
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "content": {
                    "application/problem+json": {
                        "schema": {
                            "$ref": "#/components/schemas/Problem"
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "403 insufficient_scope, role_forbidden, ip_not_allowed, write_disabled, connection_degraded, capability_locked",
                "content": {
                    "application/problem+json": {
                        "schema": {
                            "$ref": "#/components/schemas/Problem"
                        }
                    }
                }
            },
...
            "Conflict": {
                "description": "409 conflict, write_target_exists, idempotency_mismatch, idempotency_in_progress",
                "content": {
                    "application/problem+json": {
                        "schema": {
                            "$ref": "#/components/schemas/Problem"
                        }
                    }
                }
            },
            "ValidationFailed": {
                "description": "422 validation_failed mit Array errors",
                "content": {
                    "application/problem+json": {
                        "schema": {
                            "$ref": "#/components/schemas/Problem"
                        }
                    }
                }
            },
            "RateLimited": {
                "description": "429 rate_limited",
                "headers": {
                    "Retry-After": {
                        "$ref": "#/components/headers/Retry-After"
                    },
                    "RateLimit-Limit": {
                        "$ref": "#/components/headers/RateLimit-Limit"
                    }
                },
                "content": {
                    "application/problem+json": {
                        "schema": {
                            "$ref": "#/components/schemas/Problem"
                        }
                    }
                }
            },
...
            "ServiceUnavailable": {
                "description": "503 api_keys_disabled (HUB_API_KEYS_ENABLED=false), Header Retry-After",
                "content": {
                    "application/problem+json": {
                        "schema": {
                            "$ref": "#/components/schemas/Problem"
                        }
                    }
                }
            }
        },
        "schemas": {
...
            "PageMeta": {
                "type": "object",
                "required": [
                    "page",
                    "per_page",
                    "total"
                ],
                "properties": {
                    "page": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "per_page": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 500
                    },
                    "total": {
                        "type": "integer"
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "request_id": {
                        "type": "string"
                    },
                    "source": {
                        "$ref": "#/components/schemas/SourceMeta"
                    }
                }
            },
...
            "Links": {
                "type": "object",
                "properties": {
                    "self": {
                        "type": "string"
                    },
                    "first": {
                        "type": "string"
                    },
                    "last": {
                        "type": "string"
                    },
                    "next": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "prev": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                }
            },
...
            "Document": {
                "type": "object",
                "description": "Dokumentmetadaten aus dem WebDAV-Spiegel. POST legt einen Upload-Antrag für den Posteingang an.",
                "required": [
                    "id",
                    "provenance"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "folder_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "path": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "filename": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "display_name": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "content_type": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "document_type": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "size_bytes": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "remote_etag": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "remote_last_modified": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "content_hash": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "content_stored": {
                        "type": "boolean"
                    },
                    "property_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "unit_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "contact_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "case_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "assignment_confidence": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "origin": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "identity_confidence": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "identity_uncertain": {
                        "type": "boolean",
                        "description": "true, wenn die Zuordnung zur externen ID unsicher ist."
                    },
                    "provenance": {
                        "$ref": "#/components/schemas/Provenance"
                    }
                }
            },
...
        }
    },
    "security": [
        {
            "apiKey": []
        }
    ]
}